OneDrive/Desktop/crown_christian_school_platform/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
django-environ==0.11.2
Faker==25.8.0
